% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eclac_code}
\alias{eclac_code}
\title{United Nations Economic Commission for Latin America and the Caribbean Body of Water Codes}
\format{
A data frame with 13 rows and 2 variables:
\describe{
\item{eclac_code}{The ECLAC code}
\item{water_body}{Water body represented by the code}
...
}
}
\source{
\url{http://www.unece.org/cefact/codesfortrade/codes_index.html/}
}
\usage{
eclac_code
}
\description{
A dataset containing the ECLAC Codes
}
\keyword{datasets}
