% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{function_code}
\alias{function_code}
\title{Function Codes}
\format{
A data frame with 14 rows and 2 variables:
\describe{
\item{funct_code}{Single digit function code.}
\item{function_type}{The type of function represented by each code.}
...
}
}
\source{
\url{http://www.unece.org/cefact/codesfortrade/codes_index.html/}
}
\usage{
function_code
}
\description{
A dataset containing the UN Function Classifier Codes
}
\keyword{datasets}
