% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{status_code}
\alias{status_code}
\title{Status Codes}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
\item{status_code}{Two character code}
\item{definition}{The status of the entry by a 2-character code e.g. whether approved by Government, by Customs, or based on a user requirement not necessarily recognised by an authority, etc. It is also intended to show the status of checking, e.g. that function indicators are not verified.}
...
}
}
\source{
\url{http://www.unece.org/cefact/codesfortrade/codes_index.html/}
}
\usage{
status_code
}
\description{
A dataset containing the UN ST Codes
}
\keyword{datasets}
